+++ lib/easypost/internal_utilities.rb
@@ -5,3 +5,4 @@ end
 
 require_relative 'utilities/json'
 require_relative 'utilities/system'
+require_relative 'utilities/static_mapper'
