+++ lib/easypost/utilities/static_mapper.rb
@@ -1,72 +1,11 @@
 # frozen_string_literal: true
 
-module EasyPost::StaticMapper
+module EasyPost::InternalUtilities::StaticMapper
   BY_PREFIX = {
-    'ak' => EasyPost::ApiKey,
-    'adr' => EasyPost::Address,
-    'bank' => EasyPost::PaymentMethod,
-    'batch' => EasyPost::Batch,
-    'brd' => EasyPost::Brand,
-    'ca' => EasyPost::CarrierAccount,
-    'card' => EasyPost::PaymentMethod,
-    'cstinfo' => EasyPost::CustomsInfo,
-    'cstitem' => EasyPost::CustomsItem,
-    'es' => EasyPost::EndShipper,
-    'evt' => EasyPost::Event,
-    'hook' => EasyPost::Webhook,
-    'ins' => EasyPost::Insurance,
-    'order' => EasyPost::Order,
-    'payload' => EasyPost::Payload,
-    'pickup' => EasyPost::Pickup,
-    'pickuprate' => EasyPost::PickupRate,
-    'pl' => EasyPost::PostageLabel,
-    'plrep' => EasyPost::Report,
-    'prcl' => EasyPost::Parcel,
-    'rate' => EasyPost::Rate,
-    'refrep' => EasyPost::Report,
-    'rfnd' => EasyPost::Refund,
-    'sf' => EasyPost::ScanForm,
-    'shp' => EasyPost::Shipment,
-    'shpinvrep' => EasyPost::Report,
-    'shprep' => EasyPost::Report,
-    'trk' => EasyPost::Tracker,
-    'trkrep' => EasyPost::Report,
-    'user' => EasyPost::User,
+    'adr' => EasyPost::Models::Address,
   }.freeze
 
   BY_TYPE = {
-    'Address' => EasyPost::Address,
-    'ApiKey' => EasyPost::ApiKey,
-    'BankAccount' => EasyPost::PaymentMethod,
-    'Batch' => EasyPost::Batch,
-    'Brand' => EasyPost::Brand,
-    'CarbonOffset' => EasyPost::CarbonOffset,
-    'CarrierAccount' => EasyPost::CarrierAccount,
-    'CreditCard' => EasyPost::PaymentMethod,
-    'CustomsInfo' => EasyPost::CustomsInfo,
-    'CustomsItem' => EasyPost::CustomsItem,
-    'EndShipper' => EasyPost::EndShipper,
-    'Event' => EasyPost::Event,
-    'Insurance' => EasyPost::Insurance,
-    'Order' => EasyPost::Order,
-    'Parcel' => EasyPost::Parcel,
-    'PaymentLogReport' => EasyPost::Report,
-    'Pickup' => EasyPost::Pickup,
-    'PickupRate' => EasyPost::PickupRate,
-    'PostageLabel' => EasyPost::PostageLabel,
-    'Rate' => EasyPost::Rate,
-    'Referral' => EasyPost::Beta::Referral,
-    'Refund' => EasyPost::Refund,
-    'RefundReport' => EasyPost::Report,
-    'Report' => EasyPost::Report,
-    'ScanForm' => EasyPost::ScanForm,
-    'Shipment' => EasyPost::Shipment,
-    'ShipmentInvoiceReport' => EasyPost::Report,
-    'ShipmentReport' => EasyPost::Report,
-    'TaxIdentifier' => EasyPost::TaxIdentifier,
-    'Tracker' => EasyPost::Tracker,
-    'TrackerReport' => EasyPost::Report,
-    'User' => EasyPost::User,
-    'Webhook' => EasyPost::Webhook,
+    'Address' => EasyPost::Models::Address,
   }.freeze
 end
